fix community ui

+++ lib/catalogue/models/food.dart
@@ -44,7 +44,7 @@ class Fields {
   String address;
   String image;
   int ratingCount;
-  int averageRating;
+  double averageRating;
 
   Fields({
     required this.name,
